+++ kernel.py
@@ -8,9 +8,9 @@ from datetime import datetime
 try:
     def main():
         if len(sys.argv) == 2:
-            if sys.argv[1] >= "1.7.1":
+            if sys.argv[1] >= "1.8.0":
                 os.system("cls" if sys.platform == "win32" else "clear")
-                print("Procyon Neo Kernel 1.7.1")
+                print("Procyon Neo Kernel 1.8.0")
                 print("..............")
                 print("Loading Kernel modules...")
                 time.sleep(0.1)
@@ -37,7 +37,7 @@ try:
                 print("Booting ProcyonCLS...")
                 time.sleep(0.2)
                 if os.path.exists("shell.py"):
-                    os.system("python3 shell.py 1.7.1")
+                    os.system("python3 shell.py 1.8.0")
                 else:
                     os.system("cls" if sys.platform == "win32" else "clear")
                     print("Kernel Panic : OS error")
@@ -51,7 +51,7 @@ try:
                 print("Technical Details : ")
                 print(" Error Code : 0x0001")
                 print(" Error Description : Incompatible version reported by Bootloader")
-                print(f" Reported {sys.argv[1]} as opposed to 1.7.1")
+                print(f" Reported {sys.argv[1]} as opposed to 1.8.0")
                 sys.exit(1)
         else:
             os.system("cls" if sys.platform == "win32" else "clear")
@@ -79,7 +79,7 @@ def reboot():
     os.execv(sys.executable, ['python3', 'start.py'])
 
 def getVersion():
-    return "1.7.1"
+    return "1.8.0"
 
 def getBuild():
     return "2024.12.07.2108"
@@ -136,10 +136,82 @@ def bsod(error, msg):
     time.sleep(5)
     sys.exit(1)
 
+def printYellow(param):
+    print(f"\033[93m{param}\033[0m")
+
+def printGreen(param):
+    print(f"\033[92m{param}\033[0m")
+
+def printBlue(param):
+    print(f"\033[94m{param}\033[0m")
+
+def printCyan(param):
+    print(f"\033[96m{param}\033[0m")
+
+def printMagenta(param):
+    print(f"\033[95m{param}\033[0m")
+
+def printRed(param):
+    print(f"\033[91m{param}\033[0m")
+
+def printWhite(param):
+    print(f"\033[97m{param}\033[0m")
+
+def printBlack(param):
+    print(f"\033[30m{param}\033[0m")
+
+def printBold(param):
+    print(f"\033[1m{param}\033[0m")
+
+def printUnderline(param):
+    print(f"\033[4m{param}\033[0m")
+
+def printInverted(param):
+    print(f"\033[7m{param}\033[0m")
+
+def printStrikethrough(param):
+    print(f"\033[9m{param}\033[0m")
+
+def printReset(param):
+    print(f"\033[0m{param}\033[0m")
+
+def printYellow(param):
+    print(f"\033[93m{param}\033[0m")
+
+def printItalic(param):
+    print(f"\033[3m{param}\033[0m")
+
+def printOverline(param):
+    print(f"\033[53m{param}\033[0m")
+
+def printBlink(param):
+    print(f"\033[5m{param}\033[0m")
+
+def printDoubleUnderline(param):
+    print(f"\033[21m{param}\033[0m")
+
+def printDoubleStrikethrough(param):
+    print(f"\033[29m{param}\033[0m")
+
+def printDoubleOverline(param):
+    print(f"\033[55m{param}\033[0m")
+
+def printDoubleBlink(param):
+    print(f"\033[6m{param}\033[0m")
+
+def printFramed(param):
+    print(f"\033[51m{param}\033[0m")
+
+def printEncircled(param):
+    print(f"\033[52m{param}\033[0m")
+
+def printRainbow(param):
+    print(f"\033[38;2;255;0;0m{param}\033[0m")
+
 def callApplication(app, isAdmin = False):
-    appResolved = "python3 " + app + ".py 1.7.1 " + str(isAdmin)
+    appResolved = "python3 " + app + ".py 1.8.0 " + str(isAdmin)
     os.system(appResolved)
 
 def callApplication3P(app, isAdmin = False):
-    appResolved = "python3 apps/" + app + ".py 1.7.1 " + str(isAdmin)
+    appResolved = "python3 apps/" + app + ".py 1.8.0 " + str(isAdmin)
     os.system(appResolved)