+++ start.py
@@ -6,14 +6,14 @@ import time
 
 try:
     os.system("cls" if sys.platform == "win32" else "clear")
-    print("Procyon Bootloader 1.7.1")
+    print("Procyon Bootloader 1.8.0")
     print("..............")
     print("Detecting ProcyonCLS Kernel...")
     if os.path.exists("kernel.py"):
         print("Kernel found!")
         print("Booting kernel...")
         time.sleep(0.5)
-        os.system("python3 kernel.py 1.7.1")
+        os.system("python3 kernel.py 1.8.0")
     else:
         print("Kernel not found!")
         print("Exiting...")