+++ CHANGELOG.md
@@ -2,7 +2,15 @@
 
 All notable changes to this project will be documented in this file.
 
-## What's new in ProcyonCLS 2025 1.7.1 Developer Preview 10:
+## What's new in ProcyonCLS 2025 1.8.0 Developer Preview 10?
+
+* Security Update - UB-20241209.
+
+* Revamped security app.
+
+* Added new importable kernel APIs.
+
+## What was new in ProcyonCLS 2025 1.7.1 Developer Preview 10?
 
 * Security Update - UB-20241207-4.